+++ node/index.ts
@@ -1,18 +1,16 @@
-import { ClientsConfig, LRUCache, Service } from "@vtex/api";
-import { Clients } from "./clients";
+import { ClientsConfig, Service, IOContext } from "@vtex/api";
+import { Clients, BiggySearchClient } from "./clients";
 import { autocomplete } from "./resolvers/autocomplete";
-import { queries } from "./resolvers/queries";
 import { search } from "./resolvers/search";
+
+const FIFTEEN_SECOND_MS = 15 * 1000;
 
 const clients: ClientsConfig<Clients> = {
   implementation: Clients,
   options: {
     default: {
       retries: 2,
-      timeout: 12000,
-    },
-    status: {
-      memoryCache: new LRUCache<string, any>({ max: 5000 }),
+      timeout: FIFTEEN_SECOND_MS,
     },
   },
 };
@@ -23,9 +21,15 @@ export default new Service({
     resolvers: {
       Query: {
         ...autocomplete,
-        ...queries,
         ...search,
       },
     },
   },
 });
+
+export interface IContext {
+  vtex: IOContext;
+  clients: {
+    biggySearch: BiggySearchClient;
+  };
+}