+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Redirect functionality.
+
 ## [0.3.0] - 2019-11-26
 
 ### Changed
@@ -44,6 +48,3 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Initial release.
-
-[0.2.0]: https://github.com/vtex-apps/search/compare/v0.1.0...v0.2.0
-[0.1.0]: https://github.com/vtex-apps/search/releases/tag/v0.1.0